requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.1.5
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
Pillow==9.0.1
psycopg2==2.8.6
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
